// client/src/modules/wgo/components/Languages/TranslationsKeys.ts
import { ITranslations } from '../../models';

export interface ITranslationLanguageKeys extends ITranslations {
  WGO_LANGUAGE_TITLE: string;
  WGO_LANGUAGE_ADMIN_TITLE_BC: string;
  WGO_LANGUAGE_CREATE_LANGUAGE_LABEL: string;
  WGO_LANGUAGE_MODIFY_LANGUAGE_LABEL: string;
  WGO_LANGUAGE_COLUMN_LOGO_LABEL: string;
  WGO_LANGUAGE_COLUMN_CODE_LABEL: string;
  WGO_LANGUAGE_COLUMN_DEFAULT_LABEL: string;
  WGO_LANGUAGE_COLUMN_ENABLED_LABEL: string;
  WGO_LANGUAGE_COLUMN_ENABLED_STATUS: string;
  WGO_LANGUAGE_COLUMN_DISABLED_STATUS: string;
  WGO_LANGUAGE_SUCCESS_EDIT_ACTION: string;
  WGO_LANGUAGE_SUCCESS_CREATE_ACTION: string;
  WGO_LANGUAGE_ERROR_NULL_FIELD: string;
}

export const TranslationsKeys = {
  WGO_LANGUAGE_TITLE: true,
  WGO_LANGUAGE_ADMIN_TITLE_BC: true,
  WGO_LANGUAGE_CREATE_LANGUAGE_LABEL: true,
  WGO_LANGUAGE_MODIFY_LANGUAGE_LABEL: true,
  WGO_LANGUAGE_COLUMN_LOGO_LABEL: true,
  WGO_LANGUAGE_COLUMN_CODE_LABEL: true,
  WGO_LANGUAGE_COLUMN_DEFAULT_LABEL: true,
  WGO_LANGUAGE_COLUMN_ENABLED_LABEL: true,
  WGO_LANGUAGE_COLUMN_ENABLED_STATUS: true,
  WGO_LANGUAGE_COLUMN_DISABLED_STATUS: true,
  WGO_LANGUAGE_SUCCESS_EDIT_ACTION: true,
  WGO_LANGUAGE_SUCCESS_CREATE_ACTION: true,
  WGO_LANGUAGE_ERROR_NULL_FIELD: true
};